+++ src/styles/components/checkbox-style.js
@@ -1,6 +1,6 @@
 const csjs = require('csjs-inject');
 
-import common from '../common-style.js';
+import { theme } from '../../theme.js';
 
 module.exports = csjs`
 
@@ -25,7 +25,7 @@ input[type=checkbox].guify-checkbox + label:before {
     vertical-align: middle;
     margin-right: 8px;
     margin-top: 1px;
-    background-color: ${common.theme.colors.background2};
+    background-color: ${theme.colors.componentBackground};
     border-radius: 0px;
     cursor: pointer;
     box-sizing: content-box;
@@ -38,29 +38,29 @@ input[type=checkbox].guify-checkbox + label:before {
 input[type=checkbox].guify-checkbox:hover + label:before {
     width: 10px;
     height: 10px;
-    background-color: ${common.theme.colors.background2hover};
-    border: solid 4px ${common.theme.colors.background2};
-}
-
-/* Focus when checked */
-input[type=checkbox]:checked.guify-checkbox:focus + label:before {
-    width: 10px;
-    height: 10px;
-    background-color: ${common.theme.colors.foreground1};
-    border: solid 4px ${common.theme.colors.background2hover};
-}
-
-/* Focus when unchecked */
-input[type=checkbox].guify-checkbox:focus + label:before {
-    background-color: ${common.theme.colors.background2hover};
+    background-color: ${theme.colors.componentBackgroundHover};
+    border: solid 4px ${theme.colors.componentBackground};
 }
 
 /* Checked style */
 input[type=checkbox]:checked.guify-checkbox + label:before {
     width: 10px;
     height: 10px;
-    background-color: ${common.theme.colors.foreground1};
-    border: solid 4px ${common.theme.colors.background2};
+    background-color: ${theme.colors.componentForeground};
+    border: solid 4px ${theme.colors.componentBackground};
+}
+
+/* Focused and checked */
+input[type=checkbox]:checked.guify-checkbox:focus + label:before {
+    width: 10px;
+    height: 10px;
+    background-color: ${theme.colors.componentForeground};
+    border: solid 4px ${theme.colors.componentBackgroundHover};
+}
+
+/* Focus and unchecked */
+input[type=checkbox].guify-checkbox:focus + label:before {
+    background-color: ${theme.colors.componentBackgroundHover};
 }
 
 `;