+++ src/styles/components/select-style.js
@@ -1,13 +1,13 @@
 const csjs = require('csjs-inject');
 
-import common from '../common-style.js';
+import { theme } from '../../theme.js';
 
 module.exports = csjs`
 
 .guify-select-dropdown {
     display: inline-block;
     position: absolute;
-    width: calc(100% - ${common.theme.sizing.labelWidth});
+    width: calc(100% - ${theme.sizing.labelWidth});
     padding-left: 1.5%;
     height: 20px;
     border: none;
@@ -17,23 +17,28 @@ module.exports = csjs`
     -o-appearance:none;
     appearance: none;
     font-family: inherit;
-    background-color: ${common.theme.colors.background2};
-    color: ${common.theme.colors.text2};
+    background-color: ${theme.colors.componentBackground};
+    color: ${theme.colors.textSecondary};
     box-sizing: border-box !important;
     -moz-box-sizing: border-box !important;
     -webkit-box-sizing: border-box !important;
-
-    &:focus {
-        outline: none;
-    }
-    &::-moz-focus-inner {
-        border: 0;
-    }
 }
+
+/* Disable default outline since we're providing our own */
+.guify-select-dropdown:focus {
+    outline: none;
+}
+.guify-select-dropdown::-moz-focus-inner {
+    border: 0;
+}
+
+
 .guify-select-dropdown:focus,
 .guify-select-dropdown:hover {
-    background-color: ${common.theme.colors.background2hover};
+    color: ${theme.colors.textHover};
+    background-color: ${theme.colors.componentForeground};
 }
+
 .guify-select-dropdown::-ms-expand {
     display:none;
 }
@@ -45,16 +50,26 @@ module.exports = csjs`
     position: absolute;
     right: 2.5%;
     z-index: 1;
+    pointer-events: none;
 }
 .guify-select-triangle--down {
     top: 11px;
-    border-top: 5px solid ${common.theme.colors.text2};
+    border-top: 5px solid ${theme.colors.textSecondary};
     border-bottom: 0px transparent;
 }
+
 .guify-select-triangle--up {
     top: 4px;
-    border-bottom: 5px solid ${common.theme.colors.text2};
+    border-bottom: 5px solid ${theme.colors.textSecondary};
     border-top: 0px transparent;
+}
+
+.guify-select-triangle--up-highlight {
+    border-bottom-color: ${theme.colors.textHover};
+}
+
+.guify-select-triangle--down-highlight {
+    border-top-color: ${theme.colors.textHover};
 }
 
 `;