+++ src/styles/toast-area-style.js
@@ -1,15 +1,15 @@
 const csjs = require('csjs-inject');
 
-import common from './common-style.js';
+import { theme } from '../theme.js';
 
 module.exports = csjs`
 
 .guify-toast-area .guify-toast-notification:nth-child(odd) {
-    background-color: ${common.theme.colors.background2};
+    background-color: ${theme.colors.panelBackground};
 }
 
 .guify-toast-area .guify-toast-notification:nth-child(even) {
-    background-color: ${common.theme.colors.background1};
+    background-color: ${theme.colors.menuBarBackground};
 }
 
 `;