+++ src/themes.js
@@ -1,29 +1,23 @@
 module.exports = {
+
     light: {
         name: "Light",
 
         colors: {
-            background1: 'rgb(227, 227, 227)',
-            background2: 'rgb(204, 204, 204)',
-            background2hover: 'rgb(190, 190, 190)',
-            foreground1: 'rgb(105, 105, 105)',
-            text1: 'rgb(36, 36, 36)',
-            text2: 'rgb(87, 87, 87)',
-        },
+            menuBarBackground: 'rgb(227, 227, 227)',
+            menuBarText: 'rgb(36, 36, 36)',
+            panelBackground: 'rgb(227, 227, 227)',
 
-        sizing: {
-            menuBarHeight: '36px',
-            menuBarContentHeight: '20px',
-            componentHeight: '20px',
-            componentSpacing: '5px',
-            labelWidth: '42%',
-        },
+            componentBackground: 'rgb(204, 204, 204)',
+            componentBackgroundHover: 'rgb(190, 190, 190)',
+            componentForeground: 'rgb(105, 105, 105)',
+            componentActive: 'rgb(36, 36, 36)',
 
-        // Optional
-        // font: {
-        //     fontURL: '//fonts.googleapis.com/css?family=Raleway',
-        //     fontFamily: `'Raleway', sans-serif`,
-        // },
+            textPrimary: 'rgb(36, 36, 36)',
+            textSecondary: 'rgb(87, 87, 87)',
+            textHover: 'rgb(204, 204, 204)',
+            textActive: 'rgb(204, 204, 204)',
+        },
 
     },
 
@@ -31,28 +25,50 @@ module.exports = {
         name: "Dark",
 
         colors: {
-            background1: 'rgb(35, 35, 35)',
-            background2: 'rgb(54, 54, 54)',
-            background2hover: 'rgb(76, 76, 76)',
-            foreground1: 'rgb(112, 112, 112)',
-            text1: 'rgb(235, 235, 235)',
-            text2: 'rgb(161, 161, 161)',
+            menuBarBackground: 'rgb(35, 35, 35)',
+            menuBarText: 'rgb(235, 235, 235)',
+            panelBackground: 'rgb(35, 35, 35)',
+
+            componentBackground: 'rgb(54, 54, 54)',
+            componentBackgroundHover: 'rgb(76, 76, 76)',
+            componentForeground: 'rgb(112, 112, 112)',
+            componentActive: 'rgb(202, 202, 202)',
+
+            textPrimary: 'rgb(235, 235, 235)',
+            textSecondary: 'rgb(181, 181, 181)',
+            textHover: 'rgb(235, 235, 235)',
+            textActive: 'rgb(54, 54, 54)',
         },
 
-        sizing: {
-            menuBarHeight: '36px',
-            menuBarContentHeight: '20px',
-            componentHeight: '20px',
-            componentSpacing: '5px',
-            labelWidth: '42%',
+    },
+
+    // Color scheme from https://metakirby5.github.io/yorha/
+    yorha: {
+        name: "YoRHa",
+
+        colors: {
+            menuBarBackground: '#CCC8B1',
+            menuBarText: '#454138',
+            panelBackground: '#CCC8B1',
+
+            componentBackground: '#BAB5A1',
+            componentBackgroundHover: '#877F6E',
+            componentForeground: '#454138',
+            componentActive: '#978F7E',
+
+            textPrimary: '#454138',
+            textSecondary: '#454138',
+            textHover: '#CCC8B1',
+            textActive: '#CCC8B1',
         },
 
-        // Optional
-        // font: {
-        //     fontURL: '//fonts.googleapis.com/css?family=Source+Sans+Pro',
-        //     fontFamily: `'Source Sans Pro', sans-serif`,
-        // },
-    }
+        //Optional
+        font: {
+            fontFamily: `helvetica, sans-serif`,
+            fontSize: '14px',
+            fontWeight: '100'
+        },
+    },
 
 
 }